wire up retry functionality on DUP > self service, necessary reorganization
