+++ frontend/pages/hosts/details/cards/Software/SelfService/SelfService.tests.tsx
@@ -33,7 +33,6 @@ const TEST_PROPS: ISoftwareSelfServiceProps = {
     self_service: false,
   },
   router: createMockRouter(),
-  onShowInstallDetails: noop,
   onShowUninstallDetails: noop,
   refetchHostDetails: noop,
   isHostDetailsPolling: false,
